requirements.txt
django
gunicorn
django-heroku
djangorestframework
django-multiselectfield